requirements.txt
asgiref==3.10.0
brotli==1.2.0
cffi==2.0.0
colorama==0.4.6
cssselect2==0.8.0
Django==4.2.26
et_xmlfile==2.0.0
fonttools==4.60.1
openpyxl==3.1.5
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.11.0
pyphen==0.17.2
python-decouple==3.8
qrcode==8.2
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.2
waitress==3.0.2
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.2.3.post1
